requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Login==0.6.3
SQLAlchemy==2.0.23

# Database drivers
psycopg2-binary==2.9.9

# Environment configuration
python-dotenv==1.0.0

# Production WSGI server
gunicorn==21.2.0

# Testing
pytest==7.4.3

# Optional: Shapely for stricter geometry validation
# Requires C++ build tools on Windows. Install separately if needed:
# pip install shapely
# Or use pre-built wheel: pip install shapely --only-binary :all:
